Expanding *Language* based concepts.

+++ Utilities/Common/DuosexagesimalBase.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Numerics;
 
 namespace AllenCopeland.Abstraction.Utilities.Common
 {
@@ -14,7 +15,7 @@ namespace AllenCopeland.Abstraction.Utilities.Common
         {
 
         }
-        public DuosexagesimalBase(ulong value)
+        public DuosexagesimalBase(BigInteger value)
             : base(controller, value)
         {
         }
@@ -23,12 +24,12 @@ namespace AllenCopeland.Abstraction.Utilities.Common
         {
         }
 
-        protected override NumericBase GetNew(ulong value)
+        protected override NumericBase GetNew(BigInteger value)
         {
             return new DuosexagesimalBase(value);
         }
 
-        public static implicit operator DuosexagesimalBase(ulong value)
+        public static implicit operator DuosexagesimalBase(BigInteger value)
         {
             return new DuosexagesimalBase(value);
         }
