+++ Utilities/Common/OctodecimalBase.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Numerics;
 
 namespace AllenCopeland.Abstraction.Utilities.Common
 {
@@ -14,7 +15,7 @@ namespace AllenCopeland.Abstraction.Utilities.Common
         {
 
         }
-        public OctodecimalBase(ulong value)
+        public OctodecimalBase(BigInteger value)
             : base(controller, value)
         {
         }
@@ -27,12 +28,12 @@ namespace AllenCopeland.Abstraction.Utilities.Common
             }
         }
 
-        protected override NumericBase GetNew(ulong value)
+        protected override NumericBase GetNew(BigInteger value)
         {
             return new OctodecimalBase(value);
         }
 
-        public static implicit operator OctodecimalBase(ulong value)
+        public static implicit operator OctodecimalBase(BigInteger value)
         {
             return new OctodecimalBase(value);
         }
